requirements.txt
# Fast Market Dashboard - Dependencies

# Core
pandas>=2.0.0
numpy>=1.24.0

# Data fetching
httpx>=0.25.0
fredapi>=0.5.0
yfinance>=0.2.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Config
pydantic>=2.0.0
python-dotenv>=1.0.0

# Dev
black>=23.0.0
ruff>=0.1.0
pytest>=7.4.0

